restructure header initialization to avoid doing it twice
