+++ plugins/core-classes/src/main/java/org/qbicc/plugin/coreclasses/BasicInitializationBasicBlockBuilder.java
@@ -28,15 +28,6 @@ public class BasicInitializationBasicBlockBuilder extends DelegatingBasicBlockBu
     }
 
     @Override
-    public ValueHandle lengthOf(ValueHandle array) {
-        ValueType arrayType = array.getValueType();
-        if (arrayType instanceof ArrayObjectType) {
-            return instanceFieldOf(array, CoreClasses.get(ctxt).getArrayLengthField());
-        }
-        return super.lengthOf(array);
-    }
-
-    @Override
     public Value new_(ClassObjectType type) {
         Value allocated = super.new_(type);
         initializeObjectHeader(CoreClasses.get(ctxt), referenceHandle(allocated), ctxt.getLiteralFactory().literalOfType(type));
